Web-Backend/algorithms/image_detection/requirements.txt
blessed==1.20.0
certifi==2023.7.22
charset-normalizer==3.3.2
gpustat==1.1.1
idna==3.4
joblib==1.3.2
numpy==1.24.4
nvidia-ml-py==12.535.133
opencv-python==4.8.1.78
Pillow==10.1.0
psutil==5.9.6
requests==2.31.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.2.0
typing_extensions==4.8.0
urllib3==2.1.0
wcwidth==0.2.10
Flask-CORS>=3.0.0